Add ClientModifiers and Logging to Client
ClientModifiers can be used to change parts of the behavior
of the Client. Currently modifiers exist for enabling a logger,
changing the request timeout, enabling request tracing and
enabling debug mode.

A logger can be injected into the client by means of a
ClientModifier. The logger should adhere to the Logger
interface.